requirements.txt
langchain>=0.1.0
langchain-google-vertexai>=1.0.0
psycopg2-binary
SQLAlchemy
langchain-community>=0.0.30
langchain-core>=0.1.40
tiktoken
google-cloud-datastore